requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.7.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9

qdrant-client==1.15.1
langchain==1.0.5


langchain-openai==1.0.2
langchain-qdrant==1.1.0
langchain-huggingface==0.1.2
PyMuPDF==1.24.7
unstructured==0.15.6
numpy==1.26.4
asyncpg
sentence-transformers